konflux/requirements.txt
PyGithub>=2.1.1
pandas>=2.0.0
matplotlib>=3.7.0 